Move testing-internal artifact to sdk-testing
