+++ exporters/inmemory/build.gradle
@@ -13,7 +13,7 @@ dependencies {
 
     annotationProcessor libraries.auto_value
 
-    testImplementation project(':opentelemetry-testing-internal')
+    testImplementation project(':opentelemetry-sdk-testing')
 
     signature "org.codehaus.mojo.signature:java18:1.0@signature"
     signature "net.sf.androidscents.signature:android-api-level-24:7.0_r2@signature"
