+++ exporters/jaeger/build.gradle
@@ -25,7 +25,7 @@ dependencies {
             libraries.testcontainers,
             libraries.okhttp
 
-    testImplementation project(':opentelemetry-testing-internal')
+    testImplementation project(':opentelemetry-sdk-testing')
 
     testRuntime "io.grpc:grpc-netty-shaded:${grpcVersion}"
