+++ sdk-extensions/tracing-incubator/build.gradle
@@ -16,7 +16,7 @@ dependencies {
 
 
     annotationProcessor libraries.auto_value
-    testImplementation project(':opentelemetry-testing-internal')
+    testImplementation project(':opentelemetry-sdk-testing')
 
     signature "net.sf.androidscents.signature:android-api-level-24:7.0_r2@signature"
 }
