+++ sdk-extensions/tracing-incubator/src/test/java/io/opentelemetry/sdk/extensions/incubator/trace/data/DelegatingSpanDataTest.java
@@ -16,7 +16,7 @@ import io.opentelemetry.api.trace.Span.Kind;
 import io.opentelemetry.api.trace.SpanId;
 import io.opentelemetry.api.trace.TraceId;
 import io.opentelemetry.api.trace.attributes.SemanticAttributes;
-import io.opentelemetry.sdk.trace.TestSpanData;
+import io.opentelemetry.sdk.testing.trace.TestSpanData;
 import io.opentelemetry.sdk.trace.data.SpanData;
 import io.opentelemetry.sdk.trace.data.SpanData.Status;
 import org.junit.jupiter.api.Test;
