+++ sdk-extensions/tracing-incubator/src/test/java/io/opentelemetry/sdk/extensions/incubator/trace/data/SpanDataBuilderTest.java
@@ -11,7 +11,7 @@ import com.google.common.testing.EqualsTester;
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.api.trace.StatusCode;
-import io.opentelemetry.sdk.trace.TestSpanData;
+import io.opentelemetry.sdk.testing.trace.TestSpanData;
 import io.opentelemetry.sdk.trace.data.SpanData;
 import io.opentelemetry.sdk.trace.data.SpanData.Status;
 import org.junit.jupiter.api.Test;
