+++ sdk/all/build.gradle
@@ -23,10 +23,10 @@ dependencies {
     testAnnotationProcessor libraries.auto_value
     testCompileOnly libraries.auto_value_annotation
 
-    testImplementation project(':opentelemetry-testing-internal')
+    testImplementation project(':opentelemetry-sdk-testing')
     testImplementation libraries.junit_pioneer
 
-    jmh(project(':opentelemetry-testing-internal')) {
+    jmh(project(':opentelemetry-sdk-testing')) {
         // JMH doesn't handle dependencies that are duplicated between the main and jmh
         // configurations properly, but luckily here it's simple enough to just exclude transitive
         // dependencies.
