+++ sdk/common/build.gradle
@@ -19,7 +19,7 @@ dependencies {
     testAnnotationProcessor libraries.auto_value
     testCompileOnly libraries.auto_value_annotation
 
-    testImplementation project(':opentelemetry-testing-internal')
+    testImplementation project(':opentelemetry-sdk-testing')
     testImplementation libraries.junit_pioneer
 
     signature "net.sf.androidscents.signature:android-api-level-24:7.0_r2@signature"
