--- testing-internal/build.gradle
@@ -1,13 +1,0 @@
-plugins {
-    id "java"
-}
-
-description = 'Internal testing utilities'
-ext.moduleName = 'io.opentelemetry.testing-internal'
-
-dependencies {
-    api project(':opentelemetry-api')
-    api project(':opentelemetry-sdk')
-
-    annotationProcessor libraries.auto_value
-}
