--- testing-internal/src/main/java/io/opentelemetry/sdk/trace/package-info.java
@@ -1,10 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-/** Testing utilities for OpenTelemetry for internal use. */
-@ParametersAreNonnullByDefault
-package io.opentelemetry.sdk.trace;
-
-import javax.annotation.ParametersAreNonnullByDefault;
